rpc: encode: ArrayFuture: panic internally
On the write side, the client should know how many elements they're
writing, so it's simpler to assert that they call `next()` the
appropriate amount of times. We were returning an option on each and
requiring the client to `unwrap()` each, which was extra boilerplate
used after every `next()` call.

The client can still inspect the `len()` if it for some reason needs to
dynamically determine when to stop writing elements.