Copy futures-aware mutex from rust-lang-nursery/futures-rs
Copied from:

https://github.com/rust-lang-nursery/futures-rs/blob/73c66412b3c26bdc08b2f8886a20f5e13847fecf/futures-util/src/lock/mutex.rs

Minor import changes to make it build.

This will be modified to use a spinlock rather than an std mutex for the
waiter list. Dynamically allocated mutexes are not really viable in
Zephyr, the intended use of this library. Spinning is appropriate for
the waiter list because we don't sleep while holding it.