encode: relax lifetimes in write_value
Because this is recursive (because Value is a recursive data type), the
future (containing the Writer) needs to be boxed at each level of
recursion. A boxed trait typicically implies static lifetime, but this
prevents calling this function in client code when the writer's lifetime
is not static. The future only needs to live as long as the writer, so
use a parameter rather than explicitly specifying static.