decode: add must_used to things containing a reader
With the next_dyn() functions which take a mutable borrow, it is
possible to drop the wrapper for an element without consuming it from
the stream. Then the array or map can be used again without having the
correct amount consumed from the reader. Try to prevent this by warning
when the wrapper is unused.