create a traditional Future for the marker byte
Allows polling for the start of a message and getting the reader back
out of the future if polling is interrupted. An async function version
of this cannot be interrupted and must wait indefinitely holding
ownership of the reader.

Useful for canceling a read loop on message boundaries.

Also create _after_marker() decode functions so upper layers can control
polling of the marker byte;