rpc: consume the message when the receiver dies
There is a drop implementation for the ArrayFuture representing the RPC
message array which sends itself back so we can finish consuming it if
the client does not. This primarily handles the case where the response
is an error, which is the 3rd of the 4 array fields, and we need to
consume the Nil in the 4th position.

But if the ValueFuture wrapping the ArrayFuture can't be sent to the
client, the ValueFuture still needs to be consumed before moving on to
consuming the next position in the array. Unfortunately we can't handle
cleanup if the client drops the ValueFuture in the middle of consuming
it, but we can handle the case where the receiver has already dropped
its receiving channel for the response.

Without this, we don't read the entire response and can get
"InvalidData" when trying to decode the next RPC message array.